refactor: replace filtering with `eth_getLogs`
Replaces filtering operations with simple logs retrieval.

The problem with using `eth_newFilter` is that we are creating filters that are stored in the memory of the RPC node, causing issues on the providers' side.

---

While working on the changes for the `get_all_erc721_transfers`, I identified several issues. To avoid expanding the scope of this change and introducing unrelated modifications, I did not address these issues here. They have been documented in #280.